refactor(synthetic): relocate density generation
